pyproject.toml
[project]
name = "ai-applications"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.4",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pinecone-client>=6.0.0",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]
